Always use a popup surface for copying
Fixes copying on KWayland (see https://github.com/bugaevc/wl-clipboard/issues/19).
It's a shame we have to do this, since other compositors I've tested
aren't as picky about the client having a focused surface, which meant
that we were able to avoid the infamous flickering at least for copying
to the regular clipboard.