Move offer mime type handling to struct offer
Now, instead of processing offerred mime types as they
appear, we will collect them into a single buffer inside
struct offer, and then analyze the whole thing when it's
ready.

This way, in the future it'd be possible to handle the case
where we see multiple offers and don't know in adnvance which
ones are the ones we need, and which we should ignore.