Move --watch to the end of the options list in the man page
We stop processing options as soon as we see --watch,
and pass the rest of the arguments to the spawned command
unmodified. Placing --watch before other options in the
usage example could be a little bit confusing, as it appeared
to suggest that we actually do accept other options after it.