Refactor bash completion scripts
This fixes https://github.com/bugaevc/wl-clipboard/issues/62

The problem was that a long option such as --seat would
be considered to be a combination of short options ending
with the letter t, such as -pnt, and completed as if it
needed a type to go after that t.

This also fixes comments and empty lines from /etc/mime.types
being retained in wl-copy --type completions.