Ignore SIGHUP in wl-copy
Running e.g “foot wl-copy foobar” or “alacritty -e wl-copy foobar”
results in the forked copy of wl-copy being terminated and “foobar” no
longer being paste:able.

The happens because it (the forked copy) receives a SIGHUP when the
terminal exits, which terminates wl-copy.

This patch fixes this by ignoring SIGHUP. Technically, this is only
needed in the forked wl-copy. But, to handle the examples above, the
handler needs to be installed *before* the parent wl-copy exits.

To avoid a race condition, ignore SIGHUP before the fork. This should
be ok since the parent wl-copy exits almost immediately after the
fork.