Change wait() calls to waitpid()
This is a more correct thing to do, considering that there could
potentially be other child processes that wl-copy itself does not know
about, due to the current process spawning them before exec'ing the
wl-copy executable. In particular, it makes some sense to spawn a
process that produces clipboard contents (piped to stdin of wl-copy) as
a child of the process that becomes wl-copy.

Fixes #156