Prevent reentrancy in wl-paste --watch
It could happen that new offer comes in while we're in the middle of
trying to receive the previous one. This leads to issues and hangs,
since the logic is not prepared to handle reentrancy at all.

Commit 89cd9f9e430a6cd8b82fa65a9c05e530e2430ffd has prevented this from
happening by immediately rejecting all but the very first offer, unless
the --watch mode is used. This makes sense, because unless the --watch
mode is used, we are indeed only interested in pasting once.

But this issue still manifested itself in the --watch mode. Fix that one
too, by not using wl_display_roundtrip() and thus not dispatching any
events in the middle of processing an offer.

Related to https://github.com/bugaevc/wl-clipboard/pull/133

Fixes https://github.com/bugaevc/wl-clipboard/issues/137