Tweak error message about missing shell
The most probable reason for this is wl-clipboard being built without
xdg-shell support, which is something we're trying to discourage, and
something that typically happens with self-built wl-clipboard binaries,
as opposed to distro-provided ones.

Instead of just adding a note (which would only be helpful if you're
actively aware of the decision to build wl-clipboard without xdg-shell),
print a specialized message explicitly pointing at missing xdg-shell at
build time as the root issue, and ask the user to rebuild wl-clipboard.